Site now opens in react!
- converted to React
- components of site now functions
