+++ my-app/src/App.css
@@ -35,7 +35,7 @@
 
 	}
 	.explainerText {
-		background-color: aliceblue;
+
 		color: #000000;
 		width: 500px;
 	}
@@ -46,5 +46,3 @@
         background-color: #fff7fb;
         border: 2px solid black;
     }
-
-
