+++ readme.md
@@ -1,9 +1,11 @@
 The code does run locally as of 6/11/23 but it won't run the API calls in the browser of course so I'll circle back to making this project a full reality once I understand server side computing a bit better.
 
 The TODOs for this project are as such:
-1. Clean up HTML for site so pre-click looks presentable // completed (6/12/23)
+1. ~~Clean up HTML for site so pre-click looks presentable~~ // completed (6/12/23)
 
-2. Convert how the site is being displayed from using HTML to using React
+2. ~~Convert how the site is being displayed from using HTML to using React~~ // completed (6/14/23)
+- Make on click buttons work again with React
+
 3. Learn React.js to find out what needs to be done to create a waiting screen and a way to display the output.
 4. Implement what I've learned in step 3 into the HTML/CSS/JS code.
